demos/requirements.txt
ivy-core
ivy-mech
ivy-vision
ivy-demo-utils
open3d